fix-修复了用户部分，订单部分，商品部分，聊天部分，登录部分前后端接口未对齐的问题。feat-重新实现了商品浏览量部分，使其能够实时更新。使用websocket，使得能够实时接收消息，同时新增聊天中传送图片的功能

+++ BackEnd/src/main/java/org/example/campusmarket/DTO/CreateConversationRequest.java
@@ -3,18 +3,21 @@ package org.example.campusmarket.DTO;
 import lombok.Data;
 import jakarta.validation.constraints.NotNull;
 
-
 @Data
 public class CreateConversationRequest {
     /** 对方用户ID */
     @NotNull(message = "对方用户ID不能为空")
-    private Integer userId; // 对应接口的userId（对方ID）
-    /** 关联订单ID */
-    private Long orderId;
-    /** 关联商品ID */
-    private Long productId;
-    /** 对方昵称 */
+    private Integer userId;
+    
+    /** 关联订单ID（可选） */
+    private Integer orderId;
+    
+    /** 关联商品ID（可选） */
+    private Integer productId;
+    
+    /** 对方昵称（可选，如果不传则从数据库查询） */
     private String partnerName;
-    /** 对方头像 */
+    
+    /** 对方头像（可选，如果不传则从数据库查询） */
     private String partnerAvatar;
 }
